+++ package.json
@@ -1,5 +1,4 @@
 {
-
   "name": "grace-shopper",
   "version": "2.0.0",
   "description": "Some boilerplate code to get you started - get shakin'!",
